chore: Convert FoD identifier fields from int to long to prevent overflow (#1091)
fix: `fcli fod`: Change id fields from `int` to `long` to avoid potential integer overflows

+++ fcli-core/fcli-fod/src/main/java/com/fortify/cli/fod/app/helper/FoDAppDescriptor.java
@@ -35,8 +35,8 @@ public class FoDAppDescriptor extends JsonNodeHolder {
     private String emailList;
     private boolean hasMicroservices;
 
-    public Map<Integer, String> attributesAsMap() {
-        Map<Integer, String> attrMap = new HashMap<>();
+    public Map<Long, String> attributesAsMap() {
+        Map<Long, String> attrMap = new HashMap<>();
         for (FoDAttributeValueDescriptor attr : attributes) {
             attrMap.put(attr.getId(), attr.getValue());
         }
